+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - PR #2996 IO Readers: Replace `cuio::device_buffer` with `rmm::device_buffer`
 - PR #3029 Update gdf_ numeric types with stdint and move to cudf namespace
 - PR #3070 Move functions.h and related source to legacy
+- PR #2951 Allow set_index to handle a list of column names
 - PR #3128 Support MultiIndex in DataFrame.join
 
 ## Bug Fixes